+++ services/netapp/mgmt/2020-11-01/netapp/models.go
@@ -1268,8 +1268,8 @@ type BackupProperties struct {
 	Size *int64 `json:"size,omitempty"`
 	// Label - Label for backup
 	Label *string `json:"label,omitempty"`
-	// BackupType - READ-ONLY; Type of backup adhoc or scheduled
-	BackupType *string `json:"backupType,omitempty"`
+	// BackupType - READ-ONLY; Type of backup Manual or Scheduled. Possible values include: 'Manual', 'Scheduled'
+	BackupType BackupType `json:"backupType,omitempty"`
 	// FailureReason - READ-ONLY; Failure reason
 	FailureReason *string `json:"failureReason,omitempty"`
 }
@@ -2143,7 +2143,7 @@ type PoolChangeRequest struct {
 type PoolPatchProperties struct {
 	// Size - Provisioned size of the pool (in bytes). Allowed values are in 4TiB chunks (value must be multiply of 4398046511104).
 	Size *int64 `json:"size,omitempty"`
-	// QosType - The qos type of the pool. Possible values include: 'Auto', 'Manual'
+	// QosType - The qos type of the pool. Possible values include: 'QosTypeAuto', 'QosTypeManual'
 	QosType QosType `json:"qosType,omitempty"`
 }
 
@@ -2161,7 +2161,7 @@ type PoolProperties struct {
 	TotalThroughputMibps *float64 `json:"totalThroughputMibps,omitempty"`
 	// UtilizedThroughputMibps - READ-ONLY; Utilized throughput of pool in Mibps
 	UtilizedThroughputMibps *float64 `json:"utilizedThroughputMibps,omitempty"`
-	// QosType - The qos type of the pool. Possible values include: 'Auto', 'Manual'
+	// QosType - The qos type of the pool. Possible values include: 'QosTypeAuto', 'QosTypeManual'
 	QosType QosType `json:"qosType,omitempty"`
 }
 
@@ -3558,7 +3558,7 @@ type VolumeProperties struct {
 	UsageThreshold *int64 `json:"usageThreshold,omitempty"`
 	// ExportPolicy - Set of export policy rules
 	ExportPolicy *VolumePropertiesExportPolicy `json:"exportPolicy,omitempty"`
-	// ProtocolTypes - Set of protocol types, default NFSv3, CIFS fro SMB protocol
+	// ProtocolTypes - Set of protocol types, default NFSv3, CIFS for SMB protocol
 	ProtocolTypes *[]string `json:"protocolTypes,omitempty"`
 	// ProvisioningState - READ-ONLY; Azure lifecycle management
 	ProvisioningState *string `json:"provisioningState,omitempty"`