+++ services/netapp/mgmt/2021-04-01/netapp/backuppolicies.go
@@ -91,6 +91,7 @@ func (client BackupPoliciesClient) CreatePreparer(ctx context.Context, resourceG
 	body.ID = nil
 	body.Name = nil
 	body.Type = nil
+	body.SystemData = nil
 	preparer := autorest.CreatePreparer(
 		autorest.AsContentType("application/json; charset=utf-8"),
 		autorest.AsPut(),