+++ services/netapp/mgmt/2021-04-01/netapp/CHANGELOG.md
@@ -1,2 +1,10 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. Backup.SystemData
+1. BackupPolicy.SystemData